refactor(schema): split the builder module into one file per builder
A pure move mirroring the schema module layout: SchemaBuilder and the
shared error/helper stay in mod.rs, each child builder gets its own file,
and re-exports keep all existing paths working.
